requirements.txt
streamlit
streamlit-authenticator
pandas
matplotlib
PyYAML
yalm
bcrypt
